Rename MemberHolds to Account::Holds.
