+++ app/views/items/place_hold_on/_item.html.erb
@@ -1,1 +1,1 @@
-<%= button_to "Place a hold", member_holds_path(item_id: item.id), class: "btn btn", remote: true, data: { disable_with: "Saving..." } %>
+<%= button_to "Place a hold", account_holds_path(item_id: item.id), class: "btn btn", remote: true, data: { disable_with: "Saving..." } %>
