+++ config/routes.rb
@@ -19,12 +19,14 @@ Rails.application.routes.draw do
     get "/", to: "home#index"
   end
 
+  namespace :account do
+    resources :holds, only: [:create]
+  end
+
   get "member/history", to: "members#history", as: "member_loan_history"
   get "/member/loans", to: "members#loans", as: "member_loans"
   delete "/member/holds/:id", to: "members#delete_hold", as: "delete_member_hold"
   post "/member/loans/:id/renew", to: "members#renew", as: "member_loans_renew"
-
-  resources :member_holds, only: [:create]
 
   namespace :volunteer do
     resources :shifts, only: [:index, :new, :create]
